ZMQ-159: use int* instead of ssize_t* for offset_ parameter in encoder read call
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@870 7451a71c-b6c8-49cc-8254-141fc5361860

+++ libzmq/bp_pgm_sender.cpp
@@ -190,7 +190,7 @@ void zmq::bp_pgm_sender_t::out_event ()
     //  note that all data has to written in one write
     if (write_pos < write_size) {
         size_t nbytes = epgm_socket.write_one_pkt_with_offset (txw_slice + write_pos,
-            write_size - write_pos, first_message_offest);
+            write_size - write_pos, (uint16_t) first_message_offest);
 
         zmq_log (2, "wrote %iB/%iB, %s(%i)\n", (int)(write_size - write_pos), (int)nbytes, __FILE__, __LINE__);
